refactor: move RoleAccordion to separate file
